02-HEROKU-LINEBOT/requirements.txt
Flask
gunicorn
requests
line-bot-sdk